+++ benchmarks/benches/measurement.rs
@@ -3,4 +3,3 @@ use criterion::{criterion_group, criterion_main};
 
 criterion_group!(benches, criterion_plumbing_benchmark);
 criterion_main!(benches);
-